Add TLS certificate documentation

+++ ga/latest/kernel/helpers/runtime/docker-server.sh
@@ -6,6 +6,7 @@ function importKeyCert() {
   local KEY_FILE="tls.key"
   local CA_FILE="ca.crt"
   local PASSWORD=$(openssl rand -base64 32 2>/dev/null)
+  local TRUSTSTORE_PASSWORD=$(openssl rand -base64 32 2>/dev/null)
   local TMP_CERT=ca-bundle-temp.crt
   local -r CRT_DELIMITER="/-----BEGIN CERTIFICATE-----/"
   local KUBE_SA_FOLDER="/var/run/secrets/kubernetes.io/serviceaccount"
@@ -23,25 +24,27 @@ function importKeyCert() {
       -out "${KEYSTORE_FILE}" \
       -password pass:"${PASSWORD}" >&/dev/null
 
+     # Since we are creating new keystore, always write new password to a file
+    sed "s|REPLACE|$PASSWORD|g" $SNIPPETS_SOURCE/keystore.xml > $SNIPPETS_TARGET_DEFAULTS/keystore.xml
 
     # Add mounted CA to the truststore
-    if [ -f "${CERT_FOLDER}/${CA_FILE}" ] && [ -f "${CERT_FOLDER}/${CA_FILE}" ]; then
+    if [ -f "${CERT_FOLDER}/${CA_FILE}" ]; then
         echo "Found mounted TLS CA certificate, adding to truststore"
         keytool -import -storetype pkcs12 -noprompt -keystore "${TRUSTSTORE_FILE}" -file "${CERT_FOLDER}/${CA_FILE}" \
-          -storepass "${PASSWORD}" -alias "service-ca" >&/dev/null    
+          -storepass "${TRUSTSTORE_PASSWORD}" -alias "service-ca" >&/dev/null    
     fi
   fi
 
   # Add kubernetes CA certificates to the truststore
   # CA bundles need to be split and added as individual certificates
-  if [ -d "${KUBE_SA_FOLDER}" ]; then
+  if [ "$SEC_IMPORT_K8S_CERTS" = "true" ] && [ -d "${KUBE_SA_FOLDER}" ]; then
     mkdir /tmp/certs
     pushd /tmp/certs >&/dev/null
     cat ${KUBE_SA_FOLDER}/*.crt >${TMP_CERT}
     csplit -s -z -f crt- "${TMP_CERT}" "${CRT_DELIMITER}" '{*}'
     for CERT_FILE in crt-*; do
       keytool -import -storetype pkcs12 -noprompt -keystore "${TRUSTSTORE_FILE}" -file "${CERT_FILE}" \
-        -storepass "${PASSWORD}" -alias "service-sa-${CERT_FILE}" >&/dev/null
+        -storepass "${TRUSTSTORE_PASSWORD}" -alias "service-sa-${CERT_FILE}" >&/dev/null
     done
     popd >&/dev/null
     rm -rf /tmp/certs
@@ -49,12 +52,12 @@ function importKeyCert() {
 
   # Add the keystore password to server configuration
   if [ ! -e $keystorePath ]; then
-    sed -i.bak "s|REPLACE|$PASSWORD|g" $SNIPPETS_SOURCE/keystore.xml
-    cp $SNIPPETS_SOURCE/keystore.xml $SNIPPETS_TARGET_DEFAULTS/keystore.xml
+    sed "s|REPLACE|$PASSWORD|g" $SNIPPETS_SOURCE/keystore.xml > $SNIPPETS_TARGET_DEFAULTS/keystore.xml
   fi
   if [ -e $TRUSTSTORE_FILE ]; then
-    sed -i.bak "s|PWD_TRUST|$PASSWORD|g" $SNIPPETS_SOURCE/truststore.xml
-    cp $SNIPPETS_SOURCE/truststore.xml $SNIPPETS_TARGET_DEFAULTS/truststore.xml
+    sed "s|PWD_TRUST|$TRUSTSTORE_PASSWORD|g" $SNIPPETS_SOURCE/truststore.xml > $SNIPPETS_TARGET_OVERRIDES/truststore.xml
+  else
+    cp $SNIPPETS_SOURCE/trustDefault.xml $SNIPPETS_TARGET_OVERRIDES/trustDefault.xml  
   fi
 }
 
@@ -84,18 +87,6 @@ SNIPPETS_TARGET_OVERRIDES=/config/configDropins/overrides
 keystorePath="$SNIPPETS_TARGET_DEFAULTS/keystore.xml"
 
 importKeyCert
-
-if [ "$SSL" != "false" ] && [ "$TLS" != "false" ]
-then
-  if [ ! -e $keystorePath ]
-  then
-    # Generate the keystore.xml
-    export KEYSTOREPWD=$(head /dev/urandom | tr -dc A-Za-z0-9 | head -c 32 ; echo '')
-    sed -i.bak "s|REPLACE|$KEYSTOREPWD|g" $SNIPPETS_SOURCE/keystore.xml
-    cp $SNIPPETS_SOURCE/keystore.xml $SNIPPETS_TARGET_DEFAULTS/keystore.xml
-  fi
-fi
-
 
 # Pass on to the real server run
 exec "$@"
